Rename Rendering and Font subdirectories
