+++ anvil/rendering.h
@@ -1,8 +1,8 @@
 #pragma once
 #include <anvil/ecs/ecs.h>
 
-#include <sprocket/graphics/Rendering/geometry_renderer.h>
-#include <sprocket/graphics/Rendering/pbr_renderer.h>
+#include <sprocket/graphics/renderers/geometry_renderer.h>
+#include <sprocket/graphics/renderers/pbr_renderer.h>
 
 #include <glm/glm.hpp>
