+++ kinematica/kinematica.m.cpp
@@ -1,5 +1,5 @@
 #include <sprocket/core.h>
-#include <sprocket/graphics/Rendering/shape_renderer.h>
+#include <sprocket/graphics/renderers/shape_renderer.h>
 
 #include <glm/glm.hpp>
