+++ sprocket/graphics/renderers/geometry_renderer.cpp
