+++ sprocket/graphics/renderers/shape_renderer.cpp
