+++ sprocket/graphics/renderers/shape_renderer.h
