+++ sprocket/graphics/renderers/skybox_renderer.cpp
