+++ sprocket/graphics/renderers/skybox_renderer.h
